+++ test/sql/functions.sql
@@ -2,6 +2,9 @@ SELECT '[1,2,3]'::vector + '[4,5,6]';
 SELECT '[3e38]'::vector + '[3e38]';
 SELECT '[1,2,3]'::vector - '[4,5,6]';
 SELECT '[-3e38]'::vector - '[3e38]';
+SELECT '[1,2,3]'::vector * '[4,5,6]';
+SELECT '[1e37]'::vector * '[1e37]';
+SELECT '[1e-37]'::vector * '[1e-37]';
 
 SELECT vector_dims('[1,2,3]');