+++ sql/vector--0.4.4--0.5.0.sql
@@ -11,3 +11,15 @@ CREATE OPERATOR * (
 	LEFTARG = vector, RIGHTARG = vector, PROCEDURE = vector_mul,
 	COMMUTATOR = *
 );
+
+CREATE FUNCTION vector_sum(double precision[]) RETURNS vector
+	AS 'MODULE_PATHNAME' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
+
+CREATE AGGREGATE sum(vector) (
+	SFUNC = vector_accum,
+	STYPE = double precision[],
+	FINALFUNC = vector_sum,
+	COMBINEFUNC = vector_combine,
+	INITCOND = '{0}',
+	PARALLEL = SAFE
+);