+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added support for parallel index builds
 - Added `l1_distance` function
 - Added element-wise multiplication for vectors
+- Added `sum` aggregate
 
 ## 0.4.4 (2023-06-12)