+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `sum` aggregate
 - Improved performance of distance functions
 - Fixed out of range results for cosine distance
+- Fixed results for NULL and NaN distances
 
 ## 0.4.4 (2023-06-12)